tests/requirements.txt
nats-py>=2.6.0
asyncio-nats-client>=0.11.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
